+++ src/flock/__init__.py
@@ -3,7 +3,7 @@
 from flock.exceptions import FlockError
 from flock.types import NodeInfo, TaskSpec, TaskStatus
 
-__version__ = "1.0.0"
+__version__ = "1.1.0"
 __all__ = [
     "FlockError",
     "NodeInfo",